Fix functional tests after yesterday's addition of notification email to donations.
Note that this functional never worked. The donation params weren't properly constructed, and a
donation wasn't being created, yet the test reported apparent success. It turns out that the failures
it induced where actually masked by the checking logic. The controller's' create action will always result
in a 200 OK regardless of whether the donation is successfully saved or not. The logical difference is
actually which template is rendered. But since the test was only checking for HTTP success, the appearance
was always that the controller was behaving expected.

The difference in controller behavior to test is, as mentioned above, which template is rendered depending
on the donation object binary state of valid/invalid. The model/domain validity requirements are handled by
the donation unit test.